+++ src/App.js
@@ -1,22 +1,17 @@
 import React from 'react';
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
-const todoList = [
-  { id: 1, title: 'Complete assignment 1' },
-  { id: 2, title: 'Read a book' },
-  { id: 3, title: 'Exercise' }
-];
 
 function App() {
   return (
     <div style={{ textAlign: 'center' }}>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map((item) => (
-          <li key={item.id}>{item.title}</li>
-        ))}
-      </ul>
+      <TodoList />
+      <AddTodoForm />
     </div>
   );
 }
 
 export default App;
+
